Move monkeypatch_method decorator to utils/misc.
